SW-3014 remove old search directory, reimport from types and services
- removed old search index.ts which was only re-exporting
- moved some types into the types folder
- updated imports
