+++ src/App.tsx
@@ -6,7 +6,7 @@ import { Redirect, Route, Switch } from 'react-router-dom';
 import hexRgb from 'hex-rgb';
 import useStateLocation from './utils/useStateLocation';
 import { DEFAULT_SEED_SEARCH_FILTERS, DEFAULT_SEED_SEARCH_SORT_ORDER } from 'src/services/SeedBankService';
-import { SearchSortOrder, SearchCriteria } from 'src/services/SearchService';
+import { SearchSortOrder, SearchCriteria } from 'src/types/Search';
 import ContactUs from 'src/components/ContactUs';
 import EditOrganization from 'src/components/EditOrganization';
 import Home from 'src/components/Home';
