+++ src/services/FacilityService.ts
@@ -2,7 +2,8 @@ import strings from 'src/strings';
 import { paths } from 'src/api/types/generated-schema';
 import { Organization } from 'src/types/Organization';
 import { Facility, FacilityType } from 'src/types/Facility';
-import SearchService, { SearchNodePayload } from './SearchService';
+import { SearchNodePayload } from 'src/types/Search';
+import SearchService from './SearchService';
 import HttpService, { Response } from './HttpService';
 import { getAllSeedBanks, getAllNurseries } from 'src/utils/organization';
